pyproject.toml
[tool.poetry]
name = "personal-web-server"
version = "0.1.0"
description = ""
authors = ["shaggy-axel <ruslan@kortechs.io>"]

packages = [
    {include = "src"}
]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1"
python-dotenv = "^0.20.0"
django-extensions = "^3.2.0"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
django-filter = "^22.1"
drf-spectacular = "^0.24.2"
gunicorn = "^20.1.0"
djangorestframework-camel-case = "^1.3.0"
pillow = "^9.3.0"
whitenoise = "^6.3.0"
django-post-office = "^3.6.3"
django-storages = "^1.13.2"
boto3 = "^1.26.50"
django-cors-headers = "^3.13.0"
django-dirtyfields = "^1.9.1"


[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
ipython = "^8.4.0"
pre-commit = "^2.20.0"
django-stubs = "^1.12.0"
djangorestframework-stubs = "^1.7.0"
mypy = "^0.982"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-django = "^4.5.2"
model-bakery = "^1.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
exclude = '''(?x)(
    .*migrations/.*\.py$
)'''
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[tool.django-stubs]
django_settings_module = "src.shaggy.settings"
ignore_missing_model_attributes = true

[tool.poetry.scripts]
shaggy = "src.manage:main"
